#650 Refactor EspionageMission handling code and move to separate file
