+++ src/main/java/org/openRealmOfStars/AI/Mission/MissionHandling.java
@@ -46,16 +46,12 @@ import org.openRealmOfStars.player.fleet.Fleet;
 import org.openRealmOfStars.player.fleet.FleetList;
 import org.openRealmOfStars.player.leader.EspionageMission;
 import org.openRealmOfStars.player.leader.Job;
-import org.openRealmOfStars.player.leader.Leader;
-import org.openRealmOfStars.player.leader.LeaderUtility;
-import org.openRealmOfStars.player.leader.Perk;
 import org.openRealmOfStars.player.leader.stats.StatType;
 import org.openRealmOfStars.player.message.Message;
 import org.openRealmOfStars.player.message.MessageType;
 import org.openRealmOfStars.player.ship.Ship;
 import org.openRealmOfStars.player.ship.ShipHullType;
 import org.openRealmOfStars.player.ship.ShipStat;
-import org.openRealmOfStars.player.tech.Tech;
 import org.openRealmOfStars.starMap.Coordinate;
 import org.openRealmOfStars.starMap.CulturePower;
 import org.openRealmOfStars.starMap.Route;
@@ -67,7 +63,6 @@ import org.openRealmOfStars.starMap.history.event.EventType;
 import org.openRealmOfStars.starMap.newsCorp.NewsData;
 import org.openRealmOfStars.starMap.newsCorp.NewsFactory;
 import org.openRealmOfStars.starMap.planet.Planet;
-import org.openRealmOfStars.starMap.planet.construction.Building;
 import org.openRealmOfStars.utilities.DiceGenerator;
 
 /**
@@ -1903,8 +1898,8 @@ public final class MissionHandling {
                 && (selectedType == null
                    || selectedType == EspionageMission.NOT_SELECTED_YET)
                 && allowedTypes.length > 0) {
-              selectedType = scoreEspionageMission(info, planet, allowedTypes,
-                  fleet, game.getStarMap());
+              selectedType = EspionageHandling.scoreEspionageMission(
+                  info, planet, allowedTypes, fleet, game.getStarMap());
               ok = true;
             }
             if (ok) {
@@ -1915,16 +1910,16 @@ public final class MissionHandling {
                   selectedType);
               if (DiceGenerator.getRandom(100) < success) {
                 // Succeed.
-                handleSuccessfulEspionage(selectedType, planet, fleet, info,
-                    game);
+                EspionageHandling.handleSuccessfulEspionage(selectedType,
+                    planet, fleet, info, game);
                 somethingHappened = true;
               }
               int caught = EspionageUtility.calculateDetectionSuccess(planet,
                   fleet, selectedType);
               if (DiceGenerator.getRandom(100) < caught) {
                 // Caught
-                handleCaughtEspionage(selectedType, planet, fleet, info,
-                    game, somethingHappened);
+                EspionageHandling.handleCaughtEspionage(selectedType, planet,
+                    fleet, info, game, somethingHappened);
                 somethingHappened = true;
               }
               if (!somethingHappened) {
@@ -1956,1206 +1951,6 @@ public final class MissionHandling {
         }
       }
     } // End Of Espionage mission
-  }
-
-  /**
-   * Get espionage risk factor for AI attitude.
-   * @param attitude Attitude
-   * @return Risk factor from 1 to 6. 6 is taking highest risks.
-   */
-  private static int getEspionageRiskFactor(final Attitude attitude) {
-    int riskFactor = 2;
-    switch (attitude) {
-      case BACKSTABBING: {
-        riskFactor = 6;
-        break;
-      }
-      case AGGRESSIVE: {
-        riskFactor = 5;
-        break;
-      }
-      case MILITARISTIC: {
-        riskFactor = 4;
-        break;
-      }
-      case EXPANSIONIST: {
-        riskFactor = 3;
-        break;
-      }
-      case DIPLOMATIC:
-      case PEACEFUL: {
-        riskFactor = 1;
-        break;
-      }
-      default:
-      case LOGICAL:
-      case SCIENTIFIC:
-      case MERCHANTICAL: {
-        riskFactor = 2;
-        break;
-      }
-    }
-    return riskFactor;
-  }
-  /**
-   * Score espionage mission
-   * @param info Realm doing espionage
-   * @param planet Planet where to do espionage
-   * @param allowedTypes Allowed espionage types
-   * @param fleet Fleet doing espionage
-   * @param starMap Full Starmap
-   * @return EspionageMission, never null.
-   */
-  private static EspionageMission scoreEspionageMission(final PlayerInfo info,
-      final Planet planet, final EspionageMission[] allowedTypes,
-      final Fleet fleet, final StarMap starMap) {
-    int[] scores = new int[allowedTypes.length];
-    int target = planet.getPlanetOwnerIndex();
-    int infoIndex = starMap.getPlayerList().getIndex(info);
-    Attitude attitude = planet.getPlanetPlayerInfo().getAiAttitude();
-    int riskFactor = getEspionageRiskFactor(attitude);
-    int sum = 0;
-    for (int i = 0; i < allowedTypes.length; i++) {
-      int success = EspionageUtility.calculateSuccess(planet, fleet,
-          allowedTypes[i]);
-      int caught = EspionageUtility.calculateDetectionSuccess(planet,
-          fleet, allowedTypes[i]);
-      if (allowedTypes[i] == EspionageMission.ASSASSIN_GOVERNOR
-          || allowedTypes[i] == EspionageMission.DEMOLISH_BUILDING) {
-        switch (attitude) {
-          case AGGRESSIVE:
-          case MILITARISTIC:
-          case BACKSTABBING: {
-            scores[i] = 50;
-            break;
-          }
-          case DIPLOMATIC:
-          case PEACEFUL: {
-            scores[i] = 0;
-            break;
-          }
-          case EXPANSIONIST: {
-            scores[i] = 40;
-            break;
-          }
-          case LOGICAL:
-          case SCIENTIFIC: {
-            scores[i] = 25;
-            break;
-          }
-          default:
-          case MERCHANTICAL: {
-            scores[i] = 10;
-            break;
-          }
-        }
-        if (info.getDiplomacy().isWar(target)) {
-          scores[i] = scores[i] + 20;
-        }
-        if (info.getDiplomacy().isTradeEmbargo(target)) {
-          scores[i] = scores[i] + 10;
-        }
-      }
-      if (allowedTypes[i] == EspionageMission.STEAL_CREDIT) {
-        switch (attitude) {
-          case AGGRESSIVE:
-          case MILITARISTIC:
-          case BACKSTABBING: {
-            scores[i] = 50;
-            break;
-          }
-          case DIPLOMATIC:
-          case PEACEFUL: {
-            scores[i] = 20;
-            break;
-          }
-          case EXPANSIONIST: {
-            scores[i] = 40;
-            break;
-          }
-          case LOGICAL:
-          case SCIENTIFIC: {
-            scores[i] = 25;
-            break;
-          }
-          default:
-          case MERCHANTICAL: {
-            scores[i] = 40;
-            break;
-          }
-        }
-        if (info.getDiplomacy().isWar(target)) {
-          scores[i] = scores[i] + 10;
-        }
-        if (info.getDiplomacy().isTradeEmbargo(target)) {
-          scores[i] = scores[i] + 20;
-        }
-      }
-      if (allowedTypes[i] == EspionageMission.STEAL_TECH) {
-        switch (attitude) {
-          case AGGRESSIVE:
-          case MILITARISTIC:
-          case EXPANSIONIST:
-          case BACKSTABBING: {
-            scores[i] = 40;
-            break;
-          }
-          case DIPLOMATIC:
-          case MERCHANTICAL:
-          case PEACEFUL: {
-            scores[i] = 30;
-            break;
-          }
-          default:
-          case LOGICAL:
-          case SCIENTIFIC: {
-            scores[i] = 50;
-            break;
-          }
-        }
-        if (info.getDiplomacy().isWar(target)) {
-          scores[i] = scores[i] + 20;
-        }
-        if (info.getDiplomacy().isTradeEmbargo(target)) {
-          scores[i] = scores[i] + 20;
-        }
-      }
-      if (allowedTypes[i] == EspionageMission.TERRORIST_ATTACK) {
-        switch (attitude) {
-          case AGGRESSIVE:
-          case MILITARISTIC:
-          case BACKSTABBING: {
-            scores[i] = 50;
-            break;
-          }
-          case DIPLOMATIC:
-          case PEACEFUL: {
-            scores[i] = 0;
-            break;
-          }
-          case EXPANSIONIST: {
-            scores[i] = 20;
-            break;
-          }
-          case LOGICAL:
-          case SCIENTIFIC: {
-            scores[i] = 15;
-            break;
-          }
-          default:
-          case MERCHANTICAL: {
-            scores[i] = 5;
-            break;
-          }
-        }
-        if (info.getDiplomacy().isWar(target)) {
-          scores[i] = scores[i] + 20;
-        }
-        if (info.getDiplomacy().isTradeEmbargo(target)) {
-          scores[i] = scores[i] + 10;
-        }
-      }
-      if (allowedTypes[i] == EspionageMission.SABOTAGE
-          || allowedTypes[i] == EspionageMission.DEMOLISH_BUILDING) {
-        switch (attitude) {
-          case AGGRESSIVE:
-          case MILITARISTIC:
-          case BACKSTABBING: {
-            scores[i] = 60;
-            break;
-          }
-          case DIPLOMATIC:
-          case PEACEFUL: {
-            scores[i] = 5;
-            break;
-          }
-          case EXPANSIONIST: {
-            scores[i] = 40;
-            break;
-          }
-          case LOGICAL:
-          case SCIENTIFIC: {
-            scores[i] = 30;
-            break;
-          }
-          default:
-          case MERCHANTICAL: {
-            scores[i] = 10;
-            break;
-          }
-        }
-        if (info.getDiplomacy().isWar(target)) {
-          scores[i] = scores[i] + 20;
-        }
-        if (info.getDiplomacy().isTradeEmbargo(target)) {
-          scores[i] = scores[i] + 10;
-        }
-      }
-      if (allowedTypes[i] == EspionageMission.FALSE_FLAG) {
-        switch (attitude) {
-          case AGGRESSIVE:
-          case MILITARISTIC:
-          case BACKSTABBING: {
-            scores[i] = 60;
-            break;
-          }
-          case DIPLOMATIC:
-          case PEACEFUL: {
-            scores[i] = 5;
-            break;
-          }
-          case EXPANSIONIST: {
-            scores[i] = 40;
-            break;
-          }
-          case LOGICAL:
-          case SCIENTIFIC: {
-            scores[i] = 30;
-            break;
-          }
-          default:
-          case MERCHANTICAL: {
-            scores[i] = 10;
-            break;
-          }
-        }
-        int difference = starMap.getMilitaryDifference(infoIndex, target,
-            info.getDiplomacy().isWar(target));
-        if (difference > 100) {
-          scores[i] = scores[i] + 30;
-        } else if (difference > 80) {
-          scores[i] = scores[i] + 20;
-        } else if (difference > 50) {
-          scores[i] = scores[i] + 10;
-        } else if (difference > 0) {
-          scores[i] = scores[i] + 0;
-        } else {
-          scores[i] = scores[i] - 15;
-        }
-        if (info.getDiplomacy().isTradeEmbargo(target)) {
-          scores[i] = scores[i] + 20;
-        }
-        if (info.getDiplomacy().isPeace(target)) {
-          scores[i] = scores[i] + 5;
-        }
-        if (info.getDiplomacy().isTradeAlliance(target)) {
-          scores[i] = scores[i] - 5;
-        }
-        if (info.getDiplomacy().isDefensivePact(target)) {
-          scores[i] = scores[i] - 10;
-        }
-        if (info.getDiplomacy().isAlliance(target)) {
-          scores[i] = scores[i] - 10;
-        }
-      }
-      if (allowedTypes[i] == EspionageMission.DEADLY_VIRUS) {
-        switch (attitude) {
-          case BACKSTABBING: {
-            scores[i] = 40;
-            break;
-          }
-          case AGGRESSIVE: {
-            scores[i] = 20;
-            break;
-          }
-          case MILITARISTIC: {
-            scores[i] = 10;
-            break;
-          }
-          default:
-          case MERCHANTICAL:
-          case LOGICAL:
-          case SCIENTIFIC:
-          case EXPANSIONIST:
-          case DIPLOMATIC:
-          case PEACEFUL: {
-            scores[i] = 0;
-            break;
-          }
-        }
-        if (info.getDiplomacy().isWar(target)) {
-          scores[i] = scores[i] + 5;
-          Mission mission = info.getMissions().getAttackMission(
-              planet.getName());
-          if (mission != null && mission.getPhase() == MissionPhase.TREKKING) {
-            scores[i] = scores[i] + 15;
-          }
-          if (mission != null && mission.getPhase() == MissionPhase.EXECUTING) {
-            scores[i] = scores[i] + 20;
-          }
-        }
-        if (info.getDiplomacy().isPeace(target)) {
-          scores[i] = scores[i] - 10;
-        }
-        if (info.getDiplomacy().isTradeAlliance(target)) {
-          scores[i] = scores[i] - 15;
-        }
-        if (info.getDiplomacy().isDefensivePact(target)) {
-          scores[i] = scores[i] - 20;
-        }
-        if (info.getDiplomacy().isAlliance(target)) {
-          scores[i] = scores[i] - 20;
-        }
-      }
-      int likingOpposite = 1;
-      if (allowedTypes[i] == EspionageMission.GAIN_TRUST) {
-        // Liking bonus is actually reversed for gain trust mission.
-        likingOpposite = -1;
-        switch (attitude) {
-          case AGGRESSIVE:
-          case MILITARISTIC: {
-            scores[i] = 0;
-            break;
-          }
-          case BACKSTABBING: {
-            scores[i] = 40;
-            break;
-          }
-          case DIPLOMATIC:
-          case PEACEFUL: {
-            scores[i] = 60;
-            break;
-          }
-          case EXPANSIONIST: {
-            scores[i] = 20;
-            break;
-          }
-          case LOGICAL:
-          case SCIENTIFIC: {
-            scores[i] = 40;
-            break;
-          }
-          default:
-          case MERCHANTICAL: {
-            scores[i] = 40;
-            break;
-          }
-        }
-        if (info.getDiplomacy().isTradeEmbargo(target)) {
-          scores[i] = scores[i] - 40;
-        }
-        if (planet.getPlanetPlayerInfo().getDiplomacy()
-            .getLiking(infoIndex) < 0) {
-          scores[i] = scores[i] + 80;
-        }
-        if (info.getDiplomacy().isWar(target)) {
-          scores[i] = 0;
-        }
-      }
-      // Adds liking for scores.
-      switch (info.getDiplomacy().getLiking(planet.getPlanetOwnerIndex())) {
-        case Diplomacy.HATE: {
-          scores[i] = scores[i] + 30 * likingOpposite;
-          break;
-        }
-        case Diplomacy.DISLIKE: {
-          scores[i] = scores[i] + 15 * likingOpposite;
-          break;
-        }
-        default:
-        case Diplomacy.NEUTRAL: {
-          scores[i] = scores[i] + 5 * likingOpposite;
-          break;
-        }
-        case Diplomacy.LIKE: {
-          scores[i] = scores[i] - 10 * likingOpposite;
-          break;
-        }
-        case Diplomacy.FRIENDS: {
-          scores[i] = scores[i] - 30 * likingOpposite;
-          break;
-        }
-      }
-      // Subtract possible caught chance adjusted by risk factor
-      scores[i] = scores[i] - (caught / riskFactor);
-      // Scores should not go above success rate.
-      if (scores[i] > success) {
-        scores[i] = success;
-      }
-      if (scores[i] < 0) {
-        scores[i] = 0;
-      }
-      sum = sum + scores[i];
-    }
-    if (sum > 0) {
-      int rand = DiceGenerator.getRandom(sum - 1);
-      int total = 0;
-      for (int i = 0; i < scores.length; i++) {
-        if (rand < total + scores[i]) {
-          return allowedTypes[i];
-        }
-        total = total + scores[i];
-      }
-    }
-    // Just making sure that null is never returned.
-    EspionageMission selectedType = allowedTypes[DiceGenerator.getRandom(
-        allowedTypes.length - 1)];
-    return selectedType;
-  }
-
-  /**
-   * Handle caughtespionage mission.
-   * @param type EspionageMission type
-   * @param planet Planet where to do espionage
-   * @param fleet Fleet who is doing the espionage
-   * @param info Realm who is doing the espionage
-   * @param game Full game information.
-   * @param espionageSucceed If true then espionage mission succeed, but now
-   *        leader will also get caught.
-   */
-  private static void handleCaughtEspionage(final EspionageMission type,
-      final Planet planet, final Fleet fleet, final PlayerInfo info,
-      final Game game, final boolean espionageSucceed) {
-    int infoIndex = game.getStarMap().getPlayerList().getIndex(info);
-    boolean war = planet.getPlanetPlayerInfo().getDiplomacy().isWar(infoIndex);
-    boolean tradeWar = planet.getPlanetPlayerInfo().getDiplomacy()
-        .isTradeEmbargo(infoIndex);
-    if (type == EspionageMission.GAIN_TRUST) {
-      DiplomacyBonusList diplomacy = planet.getPlanetPlayerInfo()
-          .getDiplomacy().getDiplomacyList(infoIndex);
-      if (diplomacy != null) {
-        diplomacy.addBonus(DiplomacyBonusType.ESPIONAGE_BORDER_CROSS,
-            planet.getPlanetPlayerInfo().getRace());
-        LeaderUtility.handleLeaderReleased(info, planet, fleet,
-            fleet.getCommander().getCallName() + " from "
-                + info.getEmpireName() + " caught by "
-                + planet.getPlanetPlayerInfo().getEmpireName() + " while doing"
-                + " espionage mission. Since it espionage mission was gaining"
-                + " trust " + fleet.getCommander().getCallName()
-                + " was released.", game.getStarMap());
-      }
-    }
-    if (type == EspionageMission.STEAL_CREDIT) {
-      DiplomacyBonusList diplomacy = planet.getPlanetPlayerInfo()
-          .getDiplomacy().getDiplomacyList(infoIndex);
-      if (diplomacy != null) {
-        diplomacy.addBonus(DiplomacyBonusType.ESPIONAGE_BORDER_CROSS,
-            planet.getPlanetPlayerInfo().getRace());
-      }
-      Attitude attitude = planet.getPlanetPlayerInfo().getAiAttitude();
-      if (attitude == Attitude.AGGRESSIVE
-          || attitude == Attitude.BACKSTABBING
-          || attitude == Attitude.MILITARISTIC
-          || attitude == Attitude.MERCHANTICAL
-          || war || tradeWar) {
-        String startText = fleet.getCommander().getCallName() + " from "
-            + info.getEmpireName() + " caught by "
-            + planet.getPlanetPlayerInfo().getEmpireName() + " while doing"
-            + " espionage mission. Main goal was steal credits. ";
-        String endText = "";
-        if (war) {
-          endText = " Execution was done because of war times.";
-        } else if (tradeWar) {
-          endText = " Execution was done because of revenge of trade war.";
-        }
-        LeaderUtility.handleLeaderKilled(info, planet, fleet,
-            startText
-            + fleet.getCommander().getCallName() + " was able to escape"
-            + " from " + planet.getPlanetPlayerInfo().getEmpireName()
-            + " execution by using massive amount of credits. ",
-            startText
-            + fleet.getCommander().getCallName() + " was executed by "
-            + planet.getPlanetPlayerInfo().getEmpireName()
-            + "." + endText, game);
-      } else {
-        String startText = fleet.getCommander().getCallName() + " from "
-            + info.getEmpireName() + " caught by "
-            + planet.getPlanetPlayerInfo().getEmpireName() + " while doing"
-            + " espionage mission. Main goal was steal credits. ";
-        LeaderUtility.handleLeaderPrison(info, planet, fleet, startText
-            + fleet.getCommander().getCallName() + " was able to escape"
-            + " from " + planet.getPlanetPlayerInfo().getEmpireName()
-            + " prison time by using massive amount of credits. ",
-            startText
-            + planet.getPlanetPlayerInfo().getEmpireName() + " decided to"
-            + " prison " + fleet.getCommander().getCallName()
-            + " because of stealing.", "Steal credits", 5, game);
-      }
-    }
-    if (type == EspionageMission.STEAL_TECH) {
-      DiplomacyBonusList diplomacy = planet.getPlanetPlayerInfo()
-          .getDiplomacy().getDiplomacyList(infoIndex);
-      if (diplomacy != null) {
-        diplomacy.addBonus(DiplomacyBonusType.ESPIONAGE_BORDER_CROSS,
-            planet.getPlanetPlayerInfo().getRace());
-      }
-      Attitude attitude = planet.getPlanetPlayerInfo().getAiAttitude();
-      if (attitude == Attitude.AGGRESSIVE
-          || attitude == Attitude.BACKSTABBING
-          || attitude == Attitude.MILITARISTIC
-          || attitude == Attitude.SCIENTIFIC
-          || war || tradeWar) {
-        String startText = fleet.getCommander().getCallName() + " from "
-            + info.getEmpireName() + " caught by "
-            + planet.getPlanetPlayerInfo().getEmpireName() + " while doing"
-            + " espionage mission. Main goal was steal technology. ";
-        String endText = "";
-        if (war) {
-          endText = " Execution was done because of war times.";
-        } else if (tradeWar) {
-          endText = " Execution was done because of revenge of trade war.";
-        }
-        LeaderUtility.handleLeaderKilled(info, planet, fleet,
-            startText
-            + fleet.getCommander().getCallName() + " was able to escape"
-            + " from " + planet.getPlanetPlayerInfo().getEmpireName()
-            + " execution by using massive amount of credits. ",
-            startText
-            + fleet.getCommander().getCallName() + " was executed by "
-            + planet.getPlanetPlayerInfo().getEmpireName()
-            + "." + endText, game);
-      } else {
-        String startText = fleet.getCommander().getCallName() + " from "
-            + info.getEmpireName() + " caught by "
-            + planet.getPlanetPlayerInfo().getEmpireName() + " while doing"
-            + " espionage mission. Main goal was steal technology. ";
-        LeaderUtility.handleLeaderPrison(info, planet, fleet, startText
-            + fleet.getCommander().getCallName() + " was able to escape"
-            + " from " + planet.getPlanetPlayerInfo().getEmpireName()
-            + " prison time by using massive amount of credits. ",
-            startText
-            + planet.getPlanetPlayerInfo().getEmpireName() + " decided to"
-            + " prison " + fleet.getCommander().getCallName()
-            + " because of stealing technology.", "Steal technology", 10,
-            game);
-      }
-    }
-    if (type == EspionageMission.SABOTAGE) {
-      DiplomacyBonusList diplomacy = planet.getPlanetPlayerInfo()
-          .getDiplomacy().getDiplomacyList(infoIndex);
-      if (diplomacy != null) {
-        diplomacy.addBonus(DiplomacyBonusType.ESPIONAGE_BORDER_CROSS,
-            planet.getPlanetPlayerInfo().getRace());
-      }
-      Attitude attitude = planet.getPlanetPlayerInfo().getAiAttitude();
-      if (attitude == Attitude.AGGRESSIVE
-          || attitude == Attitude.BACKSTABBING
-          || attitude == Attitude.MILITARISTIC
-          || war || tradeWar) {
-        String startText = fleet.getCommander().getCallName() + " from "
-            + info.getEmpireName() + " caught by "
-            + planet.getPlanetPlayerInfo().getEmpireName() + " while doing"
-            + " espionage mission. Main goal was sabotage. ";
-        String endText = "";
-        if (war) {
-          endText = " Execution was done because of war times.";
-        } else if (tradeWar) {
-          endText = " Execution was done because of revenge of trade war.";
-        }
-        LeaderUtility.handleLeaderKilled(info, planet, fleet,
-            startText
-            + fleet.getCommander().getCallName() + " was able to escape"
-            + " from " + planet.getPlanetPlayerInfo().getEmpireName()
-            + " execution by using massive amount of credits.",
-            startText
-            + fleet.getCommander().getCallName() + " was executed by "
-            + planet.getPlanetPlayerInfo().getEmpireName()
-            + "." + endText, game);
-      } else {
-        String startText = fleet.getCommander().getCallName() + " from "
-            + info.getEmpireName() + " caught by "
-            + planet.getPlanetPlayerInfo().getEmpireName() + " while doing"
-            + " espionage mission. Main goal was sabotage. ";
-        LeaderUtility.handleLeaderPrison(info, planet, fleet, startText
-            + fleet.getCommander().getCallName() + " was able to escape"
-            + " from " + planet.getPlanetPlayerInfo().getEmpireName()
-            + " prison time by using massive amount of credits. ",
-            startText
-            + planet.getPlanetPlayerInfo().getEmpireName() + " decided to"
-            + " prison " + fleet.getCommander().getCallName()
-            + " because of sabotage.", "Sabotage", 8, game);
-      }
-    }
-    if (type == EspionageMission.DEMOLISH_BUILDING) {
-      DiplomacyBonusList diplomacy = planet.getPlanetPlayerInfo()
-          .getDiplomacy().getDiplomacyList(infoIndex);
-      if (diplomacy != null) {
-        diplomacy.addBonus(DiplomacyBonusType.ESPIONAGE_BORDER_CROSS,
-            planet.getPlanetPlayerInfo().getRace());
-      }
-      Attitude attitude = planet.getPlanetPlayerInfo().getAiAttitude();
-      if (attitude == Attitude.AGGRESSIVE
-          || attitude == Attitude.BACKSTABBING
-          || attitude == Attitude.MILITARISTIC
-          || war || tradeWar) {
-        String startText = fleet.getCommander().getCallName() + " from "
-            + info.getEmpireName() + " caught by "
-            + planet.getPlanetPlayerInfo().getEmpireName() + " while doing"
-            + " espionage mission. Main goal was demolish building. ";
-        String endText = "";
-        if (war) {
-          endText = " Execution was done because of war times.";
-        } else if (tradeWar) {
-          endText = " Execution was done because of revenge of trade war.";
-        }
-        LeaderUtility.handleLeaderKilled(info, planet, fleet,
-            startText
-            + fleet.getCommander().getCallName() + " was able to escape"
-            + " from " + planet.getPlanetPlayerInfo().getEmpireName()
-            + " execution by using massive amount of credits. ",
-            startText
-            + fleet.getCommander().getCallName() + " was executed by "
-            + planet.getPlanetPlayerInfo().getEmpireName()
-            + "." + endText, game);
-      } else {
-        String startText = fleet.getCommander().getCallName() + " from "
-            + info.getEmpireName() + " caught by "
-            + planet.getPlanetPlayerInfo().getEmpireName() + " while doing"
-            + " espionage mission. Main goal was demolish building. ";
-        LeaderUtility.handleLeaderPrison(info, planet, fleet, startText
-            + fleet.getCommander().getCallName() + " was able to escape"
-            + " from " + planet.getPlanetPlayerInfo().getEmpireName()
-            + " prison time by using massive amount of credits. ",
-            startText
-            + planet.getPlanetPlayerInfo().getEmpireName() + " decided to"
-            + " prison " + fleet.getCommander().getCallName()
-            + " because of demolish building.", "Demolish building", 10, game);
-      }
-    }
-    if (type == EspionageMission.ASSASSIN_GOVERNOR) {
-      DiplomacyBonusList diplomacy = planet.getPlanetPlayerInfo()
-          .getDiplomacy().getDiplomacyList(infoIndex);
-      if (diplomacy != null) {
-        diplomacy.addBonus(DiplomacyBonusType.ESPIONAGE_BORDER_CROSS,
-            planet.getPlanetPlayerInfo().getRace());
-      }
-      Attitude attitude = planet.getPlanetPlayerInfo().getAiAttitude();
-      if (attitude == Attitude.PEACEFUL
-          && !war && !tradeWar) {
-        LeaderUtility.handleLeaderReleased(info, planet, fleet,
-            fleet.getCommander().getCallName() + " from "
-                + info.getEmpireName() + " caught by "
-                + planet.getPlanetPlayerInfo().getEmpireName() + " while doing"
-                + " espionage mission. Main goal was assassin governor. "
-                + planet.getPlanetPlayerInfo().getEmpireName() + " decided to"
-                + " release " + fleet.getCommander().getCallName()
-                + ".", game.getStarMap());
-      } else {
-        String startText = fleet.getCommander().getCallName() + " from "
-            + info.getEmpireName() + " caught by "
-            + planet.getPlanetPlayerInfo().getEmpireName() + " while doing"
-            + " espionage mission. Main goal was assassin governor. ";
-        String endText = "";
-        if (war) {
-          endText = " Execution was done because of war times.";
-        } else if (tradeWar) {
-          endText = " Execution was done because of revenge of trade war.";
-        }
-        LeaderUtility.handleLeaderKilled(info, planet, fleet,
-            startText
-            + fleet.getCommander().getCallName() + " was able to escape"
-            + " from " + planet.getPlanetPlayerInfo().getEmpireName()
-            + " execution by using massive amount of credits.",
-            startText
-            + fleet.getCommander().getCallName() + " was executed by "
-            + planet.getPlanetPlayerInfo().getEmpireName()
-            + "." + endText, game);
-      }
-    }
-    if (type == EspionageMission.TERRORIST_ATTACK) {
-      DiplomacyBonusList diplomacy = planet.getPlanetPlayerInfo()
-          .getDiplomacy().getDiplomacyList(infoIndex);
-      if (diplomacy != null) {
-        diplomacy.addBonus(DiplomacyBonusType.ESPIONAGE_BORDER_CROSS,
-            planet.getPlanetPlayerInfo().getRace());
-      }
-      String startText = fleet.getCommander().getCallName() + " from "
-          + info.getEmpireName() + " caught by "
-          + planet.getPlanetPlayerInfo().getEmpireName() + " while doing"
-          + " espionage mission. Main goal was terrorist attack on planet. ";
-      String endText = "";
-      if (war) {
-        endText = " Execution was done because of war times.";
-      } else if (tradeWar) {
-        endText = " Execution was done because of revenge of trade war.";
-      }
-      LeaderUtility.handleLeaderKilled(info, planet, fleet,
-          startText
-          + fleet.getCommander().getCallName() + " was able to escape"
-          + " from " + planet.getPlanetPlayerInfo().getEmpireName()
-          + " execution by using massive amount of credits.",
-          startText
-          + fleet.getCommander().getCallName() + " was executed by "
-          + planet.getPlanetPlayerInfo().getEmpireName()
-          + "." + endText, game);
-      if (!info.getDiplomacy().isWar(planet.getPlanetOwnerIndex())) {
-        DiplomaticTrade trade = new DiplomaticTrade(game.getStarMap(),
-            game.getPlayers().getIndex(info),
-            game.getPlayers().getIndex(planet.getPlanetPlayerInfo()));
-        trade.generateEqualTrade(NegotiationType.WAR);
-        boolean casusBelli = info.getDiplomacy().hasCasusBelli(
-            game.getPlayers().getIndex(planet.getPlanetPlayerInfo()));
-        StarMapUtilities.addWarDeclatingReputation(game.getStarMap(), info,
-            planet.getPlanetPlayerInfo());
-        NewsData newsData = NewsFactory.makeWarNews(info,
-            planet.getPlanetPlayerInfo(), planet, game.getStarMap(),
-            casusBelli);
-        game.getStarMap().getNewsCorpData().addNews(newsData);
-        game.getStarMap().getHistory().addEvent(
-            NewsFactory.makeDiplomaticEvent(planet, newsData));
-        trade.doTrades();
-        PlayerInfo defender = planet.getPlanetPlayerInfo();
-        String[] list = defender.getDiplomacy().activateDefensivePact(
-            game.getStarMap(), info);
-        if (list != null) {
-          game.getStarMap().getNewsCorpData().addNews(
-              NewsFactory.makeDefensiveActivation(info, list));
-        }
-      }
-    }
-    if (type == EspionageMission.FALSE_FLAG) {
-      DiplomacyBonusList diplomacy = planet.getPlanetPlayerInfo()
-          .getDiplomacy().getDiplomacyList(infoIndex);
-      if (diplomacy != null) {
-        diplomacy.addBonus(DiplomacyBonusType.ESPIONAGE_BORDER_CROSS,
-            planet.getPlanetPlayerInfo().getRace());
-      }
-      Ship ship = null;
-      String startText = fleet.getCommander().getCallName() + " from "
-          + info.getEmpireName() + " caught by "
-          + planet.getPlanetPlayerInfo().getEmpireName() + " while doing"
-          + " espionage mission. Main goal was false flag mission on planet.";
-      if (!espionageSucceed) {
-        ship = fleet.getShipForFalseFlag();
-        startText = startText + " " + ship.getName()
-          + " was destroyed during mission.";
-      }
-      startText = startText + " Eventually "
-          + fleet.getCommander().getCallName() + " was caught by "
-          + planet.getPlanetPlayerInfo().getEmpireName() + ". ";
-      String endText = "";
-      int index = game.getStarMap().getPlayerList().getIndex(info);
-      boolean casusBelli = planet.getPlanetPlayerInfo().getDiplomacy()
-          .hasCasusBelli(index);
-      if (casusBelli) {
-        endText = " Execution was done because of spy operation"
-            + " was considered as act of war.";
-      } else if (tradeWar) {
-        endText = " Execution was done because of revenge of trade war.";
-      } else {
-        endText = " Execution was done because of acts of conspiracy.";
-      }
-      LeaderUtility.handleLeaderKilled(info, planet, fleet,
-          startText
-          + fleet.getCommander().getCallName() + " was able to escape"
-          + " from " + planet.getPlanetPlayerInfo().getEmpireName()
-          + " execution by using massive amount of credits.",
-          startText
-          + fleet.getCommander().getCallName() + " was executed by "
-          + planet.getPlanetPlayerInfo().getEmpireName()
-          + "." + endText, game);
-      if (casusBelli) {
-        // War will be made if defender has casus belli against attack
-        // Attacker will any way declare it, so that reputation
-        // drop goes to attacker.
-        DiplomaticTrade trade = new DiplomaticTrade(game.getStarMap(),
-            game.getPlayers().getIndex(info),
-            game.getPlayers().getIndex(planet.getPlanetPlayerInfo()));
-        casusBelli = info.getDiplomacy().hasCasusBelli(
-            game.getPlayers().getIndex(planet.getPlanetPlayerInfo()));
-        trade.generateEqualTrade(NegotiationType.WAR);
-          game.getPlayers().getIndex(planet.getPlanetPlayerInfo());
-        StarMapUtilities.addWarDeclatingReputation(game.getStarMap(), info,
-            planet.getPlanetPlayerInfo());
-        NewsData newsData = NewsFactory.makeWarNews(info,
-            planet.getPlanetPlayerInfo(), planet, game.getStarMap(),
-            casusBelli);
-        game.getStarMap().getNewsCorpData().addNews(newsData);
-        game.getStarMap().getHistory().addEvent(NewsFactory.makeDiplomaticEvent(
-          planet, newsData));
-        trade.doTrades();
-        PlayerInfo defender = planet.getPlanetPlayerInfo();
-        String[] list = defender.getDiplomacy().activateDefensivePact(
-            game.getStarMap(), info);
-        if (list != null) {
-          game.getStarMap().getNewsCorpData().addNews(
-              NewsFactory.makeDefensiveActivation(info, list));
-        }
-      }
-      if (ship != null) {
-        fleet.removeShip(ship);
-        info.getFleets().recalculateList();
-      }
-      if (fleet.getCommander() != null && fleet.getNumberOfShip() == 0) {
-        // Fleet's last ship was destroyed but commander
-        // escaped from execution.
-        fleet.getCommander().setJob(Job.UNASSIGNED);
-      }
-    }
-    if (type == EspionageMission.DEADLY_VIRUS) {
-      DiplomacyBonusList diplomacy = planet.getPlanetPlayerInfo()
-          .getDiplomacy().getDiplomacyList(infoIndex);
-      if (diplomacy != null) {
-        diplomacy.addBonus(DiplomacyBonusType.ESPIONAGE_BORDER_CROSS,
-            planet.getPlanetPlayerInfo().getRace());
-      }
-      String startText = fleet.getCommander().getCallName() + " from "
-          + info.getEmpireName() + " caught by "
-          + planet.getPlanetPlayerInfo().getEmpireName() + " while doing"
-          + " espionage mission. Main goal was spreading virus on planet. ";
-      String endText = "";
-      if (war) {
-        endText = " Execution was done because of war times.";
-      } else if (tradeWar) {
-        endText = " Execution was done because of revenge of trade war.";
-      } else {
-        endText = " Spreading deadly virus was considered as attack against "
-            + planet.getPlanetPlayerInfo().getEmpireName() + ".";
-      }
-      LeaderUtility.handleLeaderKilled(info, planet, fleet,
-          startText
-          + fleet.getCommander().getCallName() + " was able to escape"
-          + " from " + planet.getPlanetPlayerInfo().getEmpireName()
-          + " execution by using massive amount of credits.",
-          startText
-          + fleet.getCommander().getCallName() + " was executed by "
-          + planet.getPlanetPlayerInfo().getEmpireName()
-          + "." + endText, game);
-      if (!info.getDiplomacy().isWar(planet.getPlanetOwnerIndex())) {
-        DiplomaticTrade trade = new DiplomaticTrade(game.getStarMap(),
-            game.getPlayers().getIndex(info),
-            game.getPlayers().getIndex(planet.getPlanetPlayerInfo()));
-        trade.generateEqualTrade(NegotiationType.WAR);
-        boolean casusBelli = info.getDiplomacy().hasCasusBelli(
-            game.getPlayers().getIndex(planet.getPlanetPlayerInfo()));
-        StarMapUtilities.addWarDeclatingReputation(game.getStarMap(), info,
-            planet.getPlanetPlayerInfo());
-        NewsData newsData = NewsFactory.makeWarNews(info,
-            planet.getPlanetPlayerInfo(), planet, game.getStarMap(),
-            casusBelli);
-        game.getStarMap().getNewsCorpData().addNews(newsData);
-        game.getStarMap().getHistory().addEvent(
-            NewsFactory.makeDiplomaticEvent(planet, newsData));
-        trade.doTrades();
-        PlayerInfo defender = planet.getPlanetPlayerInfo();
-        String[] list = defender.getDiplomacy().activateDefensivePact(
-            game.getStarMap(), info);
-        if (list != null) {
-          game.getStarMap().getNewsCorpData().addNews(
-              NewsFactory.makeDefensiveActivation(info, list));
-        }
-      }
-    }
-
-  }
-
-  /**
-   * Handle successful espionage mission.
-   * @param type EspionageMission type
-   * @param planet Planet where to do espionage
-   * @param fleet Fleet who is doing the espionage
-   * @param info Realm who is doing the espionage
-   * @param game Full game information.
-   */
-  private static void handleSuccessfulEspionage(final EspionageMission type,
-      final Planet planet, final Fleet fleet, final PlayerInfo info,
-      final Game game) {
-    int infoIndex = game.getStarMap().getPlayerList().getIndex(info);
-    if (type == EspionageMission.GAIN_TRUST) {
-      DiplomacyBonusList diplomacy = planet.getPlanetPlayerInfo()
-          .getDiplomacy().getDiplomacyList(infoIndex);
-      if (diplomacy != null) {
-        diplomacy.addBonus(DiplomacyBonusType.DIPLOMATIC_TRADE,
-            planet.getPlanetPlayerInfo().getRace());
-        Message msg = new Message(MessageType.LEADER,
-            fleet.getCommander().getCallName() + " from "
-            + info.getEmpireName() + " has gained trust against "
-            + planet.getPlanetPlayerInfo().getEmpireName() + ". This was "
-            + "gained via espionage mission.",
-            Icons.getIconByName(Icons.ICON_SPY_GOGGLES));
-        msg.setCoordinate(planet.getCoordinate());
-        msg.setMatchByString(fleet.getCommander().getName());
-        info.getMsgList().addUpcomingMessage(msg);
-        fleet.getCommander().setExperience(
-            fleet.getCommander().getExperience() + type.getExperienceReward());
-        game.getStarMap().getHistory().addEvent(NewsFactory.makeLeaderEvent(
-            fleet.getCommander(), info, game.getStarMap(), msg.getMessage()));
-      }
-    }
-    if (type == EspionageMission.STEAL_CREDIT) {
-      int totalCredits = planet.getPlanetPlayerInfo().getTotalCredits();
-      if (totalCredits > 0) {
-        int stolen = totalCredits / 10;
-        // Planet full value affects how much credit is stolen.
-        int value = planet.getFullLevel();
-        stolen = stolen * 100 / value;
-        if (stolen == 0) {
-          stolen = 1;
-        }
-        planet.getPlanetPlayerInfo().setTotalCredits(totalCredits - stolen);
-        info.setTotalCredits(info.getTotalCredits() + stolen);
-        Message msg = new Message(MessageType.LEADER,
-            fleet.getCommander().getCallName()  + " from "
-            + info.getEmpireName() + " has stolen " + stolen
-            + " credits from " + planet.getPlanetPlayerInfo().getEmpireName()
-            + ". This was gained via espionage mission.",
-            Icons.getIconByName(Icons.ICON_SPY_GOGGLES));
-        msg.setCoordinate(planet.getCoordinate());
-        msg.setMatchByString(fleet.getCommander().getName());
-        info.getMsgList().addUpcomingMessage(msg);
-        fleet.getCommander().setExperience(
-            fleet.getCommander().getExperience()
-            + type.getExperienceReward());
-        game.getStarMap().getHistory().addEvent(NewsFactory.makeLeaderEvent(
-            fleet.getCommander(), info, game.getStarMap(), msg.getMessage()));
-      }
-    }
-    if (type == EspionageMission.STEAL_TECH) {
-      Tech[] stealableTechs = EspionageUtility.getStealableTech(planet, info);
-      if (stealableTechs.length > 0) {
-        int index = DiceGenerator.getRandom(0, stealableTechs.length - 1);
-        Tech tech = stealableTechs[index];
-        info.getTechList().addTech(stealableTechs[index]);
-        Message msg = new Message(MessageType.LEADER,
-            fleet.getCommander().getCallName() + " from "
-              + info.getEmpireName() + " has stolen "
-              + tech.getName() + " technology from "
-              + planet.getPlanetPlayerInfo().getEmpireName()
-              + ". This was gained via espionage mission.",
-              Icons.getIconByName(Icons.ICON_SPY_GOGGLES));
-        msg.setCoordinate(planet.getCoordinate());
-        msg.setMatchByString(fleet.getCommander().getName());
-        info.getMsgList().addUpcomingMessage(msg);
-        fleet.getCommander().setExperience(
-            fleet.getCommander().getExperience() + type.getExperienceReward());
-        game.getStarMap().getHistory().addEvent(NewsFactory.makeLeaderEvent(
-            fleet.getCommander(), info, game.getStarMap(), msg.getMessage()));
-      }
-    }
-    if (type == EspionageMission.SABOTAGE) {
-      planet.setProdResource(planet.getProdResource() / 2);
-      Message msg = new Message(MessageType.LEADER,
-          fleet.getCommander().getCallName() + " from "
-            + info.getEmpireName() + " sabotage "
-            + planet.getUnderConstruction().getName() + " project "
-            + " at planet " + planet.getName() + ". Planet is owned by "
-            + planet.getPlanetPlayerInfo().getEmpireName()
-            + ". This was done via espionage mission.",
-            Icons.getIconByName(Icons.ICON_SPY_GOGGLES));
-      msg.setCoordinate(planet.getCoordinate());
-      msg.setMatchByString(fleet.getCommander().getName());
-      info.getMsgList().addUpcomingMessage(msg);
-      fleet.getCommander().setExperience(
-          fleet.getCommander().getExperience() + type.getExperienceReward());
-      game.getStarMap().getHistory().addEvent(NewsFactory.makeLeaderEvent(
-          fleet.getCommander(), info, game.getStarMap(), msg.getMessage()));
-    }
-    if (type == EspionageMission.DEMOLISH_BUILDING) {
-      int index = DiceGenerator.getRandom(0,
-          planet.getBuildingList().length - 1);
-      Building building = planet.getBuildingList()[index];
-      planet.removeBuilding(building);
-      Message msg = new Message(MessageType.LEADER,
-          fleet.getCommander().getCallName() + " from "
-            + info.getEmpireName() + " demolish "
-            + building.getName() + " at planet "
-            + planet.getName() + ". Planet is owned by "
-            + planet.getPlanetPlayerInfo().getEmpireName()
-            + ". This was done via espionage mission.",
-            Icons.getIconByName(Icons.ICON_SPY_GOGGLES));
-      msg.setCoordinate(planet.getCoordinate());
-      msg.setMatchByString(fleet.getCommander().getName());
-      info.getMsgList().addUpcomingMessage(msg);
-      msg = msg.copy();
-      msg.setMatchByString(planet.getName());
-      planet.getPlanetPlayerInfo().getMsgList().addUpcomingMessage(msg);
-      NewsData news = NewsFactory.makeBuildingDestroyedNews(planet, building,
-          "");
-      if (planet.getUnderConstruction() instanceof Ship
-          && building.getName().equals("Space port")) {
-        planet.setUnderConstruction(null);
-        Message msgPort = new Message(MessageType.PLANETARY,
-            "Space port from "
-              + planet.getName() + " was destroyed and ships cannot be"
-                  + " built any more. ",
-              Icons.getIconByName(Icons.ICON_HULL_TECH));
-        msgPort.setCoordinate(planet.getCoordinate());
-        msgPort.setMatchByString(planet.getName());
-        info.getMsgList().addUpcomingMessage(msgPort);
-      }
-      if (game.getStarMap().hasHumanMet(planet.getPlanetPlayerInfo())
-          || game.getStarMap().hasHumanMet(info)) {
-        game.getStarMap().getNewsCorpData().addNews(news);
-      }
-      fleet.getCommander().setExperience(
-          fleet.getCommander().getExperience() + type.getExperienceReward());
-      game.getStarMap().getHistory().addEvent(NewsFactory.makeLeaderEvent(
-          fleet.getCommander(), info, game.getStarMap(), msg.getMessage()));
-    }
-    if (type == EspionageMission.TERRORIST_ATTACK) {
-      int index = DiceGenerator.getRandom(0,
-          planet.getBuildingList().length - 1);
-      Building building = planet.getBuildingList()[index];
-      planet.removeBuilding(building);
-      if (planet.getUnderConstruction() instanceof Ship
-          && building.getName().equals("Space port")) {
-        planet.setUnderConstruction(null);
-        Message msgPort = new Message(MessageType.PLANETARY,
-            "Space port from "
-              + planet.getName() + " was destroyed and ships cannot be"
-                  + " built any more. ",
-              Icons.getIconByName(Icons.ICON_HULL_TECH));
-        msgPort.setCoordinate(planet.getCoordinate());
-        msgPort.setMatchByString(planet.getName());
-        info.getMsgList().addUpcomingMessage(msgPort);
-      }
-      String killedTxt = "";
-      if (planet.getTotalPopulation() > 1) {
-        planet.killOneWorker();
-        killedTxt = "Also population was killed during terrorist attack. ";
-      }
-      Message msg = new Message(MessageType.LEADER,
-          fleet.getCommander().getCallName() + " from "
-            + info.getEmpireName() + " destroys "
-            + building.getName() + " at planet "
-            + planet.getName() + ". " + killedTxt + "Planet is owned by "
-            + planet.getPlanetPlayerInfo().getEmpireName()
-            + ". This was done via espionage mission.",
-            Icons.getIconByName(Icons.ICON_SPY_GOGGLES));
-      msg.setCoordinate(planet.getCoordinate());
-      msg.setMatchByString(fleet.getCommander().getName());
-      info.getMsgList().addUpcomingMessage(msg);
-      msg = msg.copy();
-      msg.setMatchByString(planet.getName());
-      planet.getPlanetPlayerInfo().getMsgList().addUpcomingMessage(msg);
-      NewsData news = NewsFactory.makeBuildingDestroyedNews(planet, building,
-          killedTxt);
-      game.getStarMap().getNewsCorpData().addNews(news);
-      fleet.getCommander().setExperience(
-          fleet.getCommander().getExperience() + type.getExperienceReward());
-      game.getStarMap().getHistory().addEvent(NewsFactory.makeLeaderEvent(
-          fleet.getCommander(), info, game.getStarMap(), msg.getMessage()));
-    }
-    if (type == EspionageMission.ASSASSIN_GOVERNOR) {
-      Leader governor = planet.getGovernor();
-      if (governor.hasPerk(Perk.WEALTHY)) {
-        governor.useWealth();
-        Message msg = new Message(MessageType.LEADER,
-            fleet.getCommander().getCallName() + " from "
-              + info.getEmpireName() + " tried to assasins "
-              + governor.getCallName() + " at planet " + planet.getName()
-              + " but failed. Planet is owned by "
-              + planet.getPlanetPlayerInfo().getEmpireName()
-              + ". This was done via espionage mission.",
-              Icons.getIconByName(Icons.ICON_SPY_GOGGLES));
-        msg.setCoordinate(planet.getCoordinate());
-        msg.setMatchByString(fleet.getCommander().getName());
-        info.getMsgList().addUpcomingMessage(msg);
-        game.getStarMap().getHistory().addEvent(NewsFactory.makeLeaderEvent(
-            fleet.getCommander(), info, game.getStarMap(), msg.getMessage()));
-        msg = new Message(MessageType.LEADER,
-            governor.getCallName() + " was tried to kill "
-              + " at planet " + planet.getName()
-              + ". Governor's expensive protection gear saved "
-              + governor.getGender().getHisHer() + " life.",
-              Icons.getIconByName(Icons.ICON_SPY_GOGGLES));
-        msg.setCoordinate(planet.getCoordinate());
-        msg.setMatchByString(governor.getName());
-        planet.getPlanetPlayerInfo().getMsgList().addUpcomingMessage(msg);
-      } else {
-        planet.setGovernor(null);
-        governor.setJob(Job.DEAD);
-        fleet.getCommander().getStats().addOne(StatType.KILLED_ANOTHER_LEADER);
-        Message msg = new Message(MessageType.LEADER,
-            fleet.getCommander().getCallName() + " from "
-              + info.getEmpireName() + " assasins "
-              + governor.getCallName() + " at planet " + planet.getName()
-              + ". Planet is owned by "
-              + planet.getPlanetPlayerInfo().getEmpireName()
-              + ". This was done via espionage mission.",
-              Icons.getIconByName(Icons.ICON_SPY_GOGGLES));
-        msg.setCoordinate(planet.getCoordinate());
-        msg.setMatchByString(fleet.getCommander().getName());
-        info.getMsgList().addUpcomingMessage(msg);
-        game.getStarMap().getHistory().addEvent(NewsFactory.makeLeaderEvent(
-            fleet.getCommander(), info, game.getStarMap(), msg.getMessage()));
-        msg = new Message(MessageType.LEADER,
-            governor.getCallName() + " is killed "
-              + " at planet " + planet.getName()
-              + ". This was probably assasination.",
-              Icons.getIconByName(Icons.ICON_SPY_GOGGLES));
-        msg.setCoordinate(planet.getCoordinate());
-        msg.setMatchByString(governor.getName());
-        planet.getPlanetPlayerInfo().getMsgList().addUpcomingMessage(msg);
-        NewsData news = NewsFactory.makeLeaderDies(governor,
-            planet.getPlanetPlayerInfo(), "assasination");
-        if (game.getStarMap().hasHumanMet(planet.getPlanetPlayerInfo())) {
-          game.getStarMap().getNewsCorpData().addNews(news);
-        }
-        fleet.getCommander().setExperience(
-            fleet.getCommander().getExperience() + type.getExperienceReward());
-      }
-    }
-    if (type == EspionageMission.FALSE_FLAG) {
-      DiplomacyBonusList diplomacy = info.getDiplomacy().getDiplomacyList(
-          planet.getPlanetOwnerIndex());
-      if (diplomacy != null) {
-        diplomacy.addBonus(DiplomacyBonusType.FALSE_FLAG, info.getRace());
-      }
-      Ship ship = fleet.getShipForFalseFlag();
-      Message msg = new Message(MessageType.LEADER,
-          fleet.getCommander().getCallName() + " from "
-          + info.getEmpireName() + " makes successfull false flag"
-          + " operation against "
-          + planet.getPlanetPlayerInfo().getEmpireName() + ". Ship called "
-          + ship.getName() + " was destroyed during the mission."
-          + " This mission happened next to " + planet.getName() + ".",
-          Icons.getIconByName(Icons.ICON_SPY_GOGGLES));
-      msg.setCoordinate(planet.getCoordinate());
-      msg.setMatchByString(fleet.getCommander().getName());
-      info.getMsgList().addUpcomingMessage(msg);
-      game.getStarMap().getHistory().addEvent(NewsFactory.makeLeaderEvent(
-          fleet.getCommander(), info, game.getStarMap(), msg.getMessage()));
-      msg = new Message(MessageType.PLANETARY, ship.getName()
-          + " was destroyed from " + info.getEmpireName() + " fleet. "
-          + info.getEmpireName() + " claim that "
-          + planet.getPlanetPlayerInfo().getEmpireName()
-          + " destroyed it. This event happened next to "
-          + planet.getName() + ". ", Icons.getIconByName(Icons.ICON_DEATH));
-      msg.setCoordinate(planet.getCoordinate());
-      msg.setMatchByString(planet.getName());
-      planet.getPlanetPlayerInfo().getMsgList().addUpcomingMessage(msg);
-      fleet.removeShip(ship);
-      String location = "escaped to another ship";
-      Leader leader = fleet.getCommander();
-      if (fleet.getCommander() != null && fleet.getNumberOfShip() == 0) {
-        // Fleet's last ship was destroyed so commander escaped to
-        // friendly planet
-        fleet.getCommander().setJob(Job.UNASSIGNED);
-        location = "escaped to closest friendly planet";
-      }
-      NewsData news = NewsFactory.makeLeaderFalseFlag(leader, info,
-          planet.getPlanetPlayerInfo(), location);
-      if (game.getStarMap().hasHumanMet(info)
-          || game.getStarMap().hasHumanMet(planet.getPlanetPlayerInfo())) {
-        game.getStarMap().getNewsCorpData().addNews(news);
-      }
-      leader.setExperience(
-          leader.getExperience() + type.getExperienceReward());
-      info.getFleets().recalculateList();
-    }
-    if (type == EspionageMission.DEADLY_VIRUS) {
-      Message msg = new Message(MessageType.LEADER,
-          fleet.getCommander().getCallName() + " from "
-            + info.getEmpireName() + " spreads "
-            + " deadly virus on " + planet.getName() + ". "
-            + "Almost every population of planet is being killed."
-            + " Planet is owned by "
-            + planet.getPlanetPlayerInfo().getEmpireName()
-            + ". This was done via espionage mission.",
-            Icons.getIconByName(Icons.ICON_SPY_GOGGLES));
-      msg.setCoordinate(planet.getCoordinate());
-      msg.setMatchByString(fleet.getCommander().getName());
-      info.getMsgList().addUpcomingMessage(msg);
-      StarMapUtilities.spreadDeadlyVirus(info, planet);
-      game.getStarMap().getNewsCorpData().addNews(
-          NewsFactory.makeDeadlyVirusNews(planet, info,
-              game.getStarMap().getStarYear()));
-      fleet.getCommander().setExperience(
-          fleet.getCommander().getExperience() + type.getExperienceReward());
-      game.getStarMap().getHistory().addEvent(NewsFactory.makeLeaderEvent(
-          fleet.getCommander(), info, game.getStarMap(), msg.getMessage()));
-    }
   }
 
   /**
